requirements.txt
numpy
pandas
pyyaml
scikit-learn
torch
joblib
matplotlib
seaborn
requests
psycopg2-binary
